task: Move and rename the MetaMissionUtil
Changes:
- Rename the MetaMissionUtil to MetaMissionRegistry to better reflect the purpose
- Move to the misison top level package
- Change visibility of MetaMission constructors

ref #1298
