+++ mars-sim-console/src/main/java/com/mars_sim/console/chat/simcommand/settlement/TradeCommand.java
@@ -19,8 +19,8 @@ import com.mars_sim.core.goods.CommerceUtil;
 import com.mars_sim.core.goods.Deal;
 import com.mars_sim.core.goods.Good;
 import com.mars_sim.core.goods.ShoppingItem;
+import com.mars_sim.core.mission.MetaMissionRegistry;
 import com.mars_sim.core.person.ai.mission.MissionType;
-import com.mars_sim.core.person.ai.mission.meta.MetaMissionUtil;
 import com.mars_sim.core.structure.Settlement;
 import com.mars_sim.core.vehicle.Vehicle;
 
@@ -86,7 +86,7 @@ public class TradeCommand extends AbstractSettlementCommand {
 	private void outputDeals(Conversation context, Settlement settlement, StructuredResponse response) {
 
 		// Find some example vehicles
-		var deliveryMeta = MetaMissionUtil.getMetaMission(MissionType.DELIVERY);
+		var deliveryMeta = MetaMissionRegistry.getMetaMission(MissionType.DELIVERY);
 		var drone = deliveryMeta.selectVehicle(settlement);
 		if (drone != null) {
 			response.appendLabeledString("Drone Range", String.format(CommandHelper.KM_FORMAT,
@@ -95,7 +95,7 @@ public class TradeCommand extends AbstractSettlementCommand {
 										drone.getBaseRange()));
 		}
 
-		var tradeMeta = MetaMissionUtil.getMetaMission(MissionType.TRADE);
+		var tradeMeta = MetaMissionRegistry.getMetaMission(MissionType.TRADE);
 		var rover = tradeMeta.selectVehicle(settlement);
 		if (rover != null) {
 			response.appendLabeledString("Rover Range", String.format(CommandHelper.KM_FORMAT,
