+++ mars-sim-core/src/main/java/com/mars_sim/core/person/ai/mission/meta/CollectRegolithMeta.java
@@ -36,7 +36,7 @@ public class CollectRegolithMeta extends AbstractMetaMission {
 	/** Starting sol for this mission to commence. */
 	private static final int MIN_STARTING_SOL = 4;
 
-	CollectRegolithMeta() {
+	public CollectRegolithMeta() {
 		super(MissionType.COLLECT_REGOLITH, 4, PREFERRED_LEADER_JOBS, PREFERRED_WORKER_JOBS);
 
 		setPreferredVehicle(VehicleType.ROVER_TYPES);
