+++ mars-sim-core/src/main/java/com/mars_sim/core/person/ai/mission/meta/MiningMeta.java
@@ -40,7 +40,7 @@ public class MiningMeta extends AbstractMetaMission {
 	/** The multiplier factor. */
 	private static final double FACTOR = 5.0;
 	
-    MiningMeta() {
+    public MiningMeta() {
     	super(MissionType.MINING, 6, PREFERRED_LEADER_JOBS, PREFERRED_WORKER_JOBS);
 
 		setPreferredVehicle(Set.of(VehicleType.CARGO_ROVER, VehicleType.EXPLORER_ROVER));
