+++ mars-sim-core/src/test/java/com/mars_sim/core/mission/predefined/TestDriveMissionTest.java
@@ -20,11 +20,11 @@ import java.util.Map;
 import com.mars_sim.core.test.MarsSimUnitTest;
 import com.mars_sim.core.map.location.LocalPosition;
 import com.mars_sim.core.mission.MetaMission;
+import com.mars_sim.core.mission.MetaMissionRegistry;
 import com.mars_sim.core.mission.MissionProject;
 import com.mars_sim.core.mission.MissionVehicleProject;
 import com.mars_sim.core.person.Person;
 import com.mars_sim.core.person.ai.mission.MissionType;
-import com.mars_sim.core.person.ai.mission.meta.MetaMissionUtil;
 import com.mars_sim.core.person.ai.task.util.Worker;
 import com.mars_sim.core.resource.ResourceUtil;
 import com.mars_sim.core.structure.Settlement;
@@ -42,7 +42,7 @@ class TestDriveMissionTest extends MarsSimUnitTest {
         buildGarage(home.getBuildingManager(), new LocalPosition(0,0), BUILDING_LENGTH);
         var rover = buildRover(home, "Rover 1", null, EXPLORER_ROVER);
         Person leader = buildPerson("Leader", home);
-        var meta = MetaMissionUtil.getMetaMission(MissionType.TEST_DRIVE);
+        var meta = MetaMissionRegistry.getMetaMission(MissionType.TEST_DRIVE);
 
         List<Person> recruits = new ArrayList<>();
         // Create other members; allow for the leader alreadsy counted
